feat: Add account deletion state and UI
Introduces a new state `accountDeleted` to track if a user's account has been permanently removed.
Updates the SuperAdminDashboard to provide clearer confirmation and consequences for restaurant deletion, including informing the user that their account will be inaccessible. This prepares for future UI changes to display an "Account Removed" message to deleted users.