feat(super-admin): Add ban/unban functionality
Introduces the ability for Super Admins to ban and unban user accounts. This is achieved by adding a 'banned' status to the subscription management. When a user is banned, their account will be deactivated and they will be unable to log in. The UI for managing user subscriptions has been updated to reflect this new status.